pyproject.toml
[project]
name = "pi-benchmark"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "transformers>=4.40",
    "torch>=2.0",
    "datasets>=2.0",
    "pandas",
    "tabulate",
    "safetensors",
    "huggingface-hub",
    "sentence-transformers",
    "onnxruntime",
    "tokenizers",
    "pyyaml>=6.0",
    "simple-defender-py",
    "scikit-learn>=1.4",
    "lightgbm>=4.0",
]

[tool.uv.sources]
simple-defender-py = { path = "../../simple-defender/simple-defender-py" }

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
